♻️ Structure entrypoint
