+++ src/common/securestorage.js
