+++ src/views/account/components/join.tsx
