+++ src/views/account/components/privacy.js
